Match up precedence with Go
Also adds the idea of a "Located" expression, which lets me see where
the error should point at.  This helps a little with ensuring we're
parsing everything, although paren matching means we capture too much in
the outer expression even if Pest doesn't match it all.

Should be fine when we're not just matching expressions, as we include
the EOF in the match.