Pre-commit shouldn't apply warnings that aren't configured
At the same time, a newer clippy has new warnings and we might as well
fix them.

The rationale here is that the warnings we've configured as an explicit
deny are those recommended as "stable" -- they shouldn't trigger new
errors when new versions of Rust are released.  But erroring on every
warning, as tempting as it may be, leads to breakage when GitHub updates
its Rust version.