Allow SourceModule to be Covariant wrt lifetimes
The requirement is that we must be able to return a reference to the
module with a strictly shorter lifetime than the owning reference.  We
can prove that by explicitly decoupling the module lifetime from the
lifetime of the owning string -- we don't actually change the code we
run, just how we explain the lifetimes to the compiler.